requirements.txt
flask
python-dotenv
chromadb
pypdf
python-docx
requests
tqdm
